[PM-20160] Address audit Issue I
Switches to 64-bit unsigned integers for reference counts, to avoid
potential overflow a 4B objects.

Removes cast to an signed reference count on applying refcount deltas,
in favour of a checked signed addition.

Note that we *are* still leaving failure here as a panic, because it
represents an unrecoverable backend error.

The remediation suggests special-casing the 'hot counter' empty node --
this commit does not do so, however a larger effort in #71 inlines these
in almost all cases, effectively resolving this point.

+++ storage/src/arena.rs
@@ -953,7 +953,7 @@ struct Node {
     /// doesn't tell us that the back-end data for `sp.root` can be uncached,
     /// since some other `Sp<U>` with the same hash could still be referencing
     /// that back-end data.
-    ref_count: u32,
+    ref_count: u64,
 }
 
 impl Node {
